+++ mobile/lib/providers/infrastructure/user.provider.dart
@@ -12,7 +12,7 @@ final userApiRepositoryProvider = Provider((ref) => UserApiRepository(ref.watch(
 final userServiceProvider = Provider(
   (ref) => UserService(
     userApiRepository: ref.watch(userApiRepositoryProvider),
-    userRepository: ref.watch(userRepositoryProvider),
+    userRepository: ref.watch(driftProvider).userRepository,
     storeService: ref.watch(storeServiceProvider),
   ),
 );
